requirements.txt
requests==2.28.2
python-dotenv==0.9.1
